[Lang] Sparse adstack heap: extract the lazy-claim / bound-reducer / split-heap methods
(publish_adstack_lazy_claim_buffers, publish_per_task_bound_count_cpu, publish_per_task_bound_count_device, ensure_per_task_float_heap_post_reducer, ensure_adstack_heap_int, ensure_adstack_heap_float, ensure_adstack_heap, publish_adstack_metadata, check_adstack_overflow) plus the eval_cmp / encode_cmp_op_for_llvm_reducer anonymous-namespace helpers from llvm_runtime_executor.cpp into a new TU llvm_adstack_lazy_claim.cpp, mirroring the SPIR-V backend's runtime/gfx/adstack_bound_reducer_launch.cpp split; pure move with no logic changes - declarations stay in llvm_runtime_executor.h
