+++ quadrants/runtime/llvm/CMakeLists.txt
@@ -4,6 +4,7 @@ add_library(llvm_runtime)
 target_sources(llvm_runtime
   PRIVATE
     llvm_runtime_executor.cpp
+    llvm_adstack_lazy_claim.cpp
     llvm_context.cpp
     snode_tree_buffer_manager.cpp
     kernel_launcher.cpp
